refactor(runtime): split prompt presets into per-scenario files
将原先堆在单个 prompts.ts 里的内置场景拆成每个场景一个文件
（presets/*.ts），types.ts 放类型、registry.ts 聚合数组与 helper，
index.ts 顶层 re-export。对外导出签名不变，逐字节验证内容一致。
