+++ packages/runtime/src/prompts/index.ts
@@ -1,1 +1,2 @@
-export * from './prompts.js';
+export * from './types.js';
+export * from './registry.js';
